Added basic wrapping system
`endow` allows a user to wrap a function and give the wrapper arbitrary attributes, including callables. Callables are then given the wrapper itself as an attribute to allow the user to call back the wrapped function. Such a system is very useful because it allows to inject arbitrary code into any function before and after its execution.